+++ package.json
@@ -14,7 +14,8 @@
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-query": "^3.39.0",
-    "recoil": "^0.7.3-alpha.2"
+    "recoil": "^0.7.3-alpha.2",
+    "recoil-persist": "^4.2.0"
   },
   "devDependencies": {
     "@types/node": "17.0.35",
