+++ pages/api/fetchPhotos.ts
@@ -1,4 +1,4 @@
-import { api } from "../utils/myApi";
+import { api } from "../../utils/myApi";
 
 export const fetchPhotos = async (pageParams: number) => {
   const { data } = await api.get(`/photos?page=${pageParams}`);
