+++ styles/list.css
@@ -1,12 +1,17 @@
 .product-container {
   display: flex;
   justify-content: center;
-
   flex-wrap: wrap;
   gap: 20px;
   padding: 50px 240px;
 }
-
+.product-img {
+  top: 0;
+  left: 0;
+  width: 200px;
+  height: 200px;
+  object-fit: cover;
+}
 .product-info {
   display: flex;
   flex-direction: column;
