+++ styles/nav.css
@@ -2,7 +2,7 @@
   width: 100%;
   height: 80px;
 
-  background: #2ac1bc;
+  background: rgb(116, 116, 238);
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.3);
 }
 
@@ -18,6 +18,7 @@
 .nav-button {
   font-weight: 500;
   font-size: 24px;
-
+  background-color: transparent;
+  border: none;
   color: #ffffff;
 }
